Version 1.3.4 Build 17
Now, the calibration command will not be sent twice within 2 seconds even if the frontend requests it. Update to print block to allow lowercase letters and to allow the same text to be sent again. Also, send the current language to the frontend. Plus a fix to updating the robot type when it changes.