tablet orientation and misc.
Changed the tablet orientation responses to give the right key relative to the camera.  Changed the robot disconnected message to only use a message from the frontend. Added the ability to send all NSLogs to a file (though left it commented out). Added a lock to read mode checking to maybe prevent some crashes, not sure this will really help.