Misc.
Now sends the turnOffCommand when the stop button is pressed in addition to setting all outputs back to original state. This way the bit buzzer will turn off even if in the middle of buzzing. Changed the battery red threshold for bit to 4.4. Changed the time when the language is being set by the system because the category names were not being changed.