Audio recordings, robot descriptions, tablet orientation
Fixed bug in iPhones in which audio recordings were always extra quiet. Apparently, when the audio session is set to .playAndRecord, iOS automatically quiets things if the microphone is too close to the speaker. Removed robot descriptions and callback text as all text is now handled by the frontend. Fixed tablet orientation bug in which the in between orientation was not being translated.